Create Bathroom_Switcher.ino
This sketch controls a 7 segment display and two relays to flip flop sets of LED's on or off.  If both of the magnetic read switches are closed, the 7 segment display will begin to count up and the LED's change to the occupied state.  If either of the switches are open the counter resets to all zero's and the lights switch to the available state.